update model adapter
