+++ evalscope/benchmarks/data_adapter.py
@@ -139,11 +139,11 @@ class DataAdapter(ABC):
                 prompt_d = self.gen_prompt(input_d=sample_d, subset_name=sub_name, few_shot_list=few_shot_data)
                 prompt_d[AnswerKeys.RAW_INPUT] = sample_d
                 res_dict[sub_name].append(prompt_d)
-
-        rnd = random.Random()
-        rnd.seed(42)
-        for k, v in res_dict.items():
-            rnd.shuffle(v)
+        # Note: for multiprocess
+        # rnd = random.Random()
+        # rnd.seed(42)
+        # for k, v in res_dict.items():
+        #     rnd.shuffle(v)
 
         return res_dict
