+++ evalscope/models/__init__.py
@@ -1,4 +1,5 @@
 # Copyright (c) Alibaba, Inc. and its affiliates.
 
+from evalscope.models.custom import *
 from evalscope.models.model import BaseModel, ChatBaseModel
 from evalscope.models.model_adapter import *
