--- evalscope/models/api/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (c) Alibaba, Inc. and its affiliates.
-
-from evalscope.models.api.openai_api import OpenaiApi
