+++ evalscope/run.py
@@ -13,7 +13,7 @@ from evalscope.benchmarks import Benchmark, BenchmarkMeta
 from evalscope.config import TaskConfig, parse_task_config
 from evalscope.constants import DEFAULT_MODEL_REVISION, DEFAULT_WORK_DIR, EvalBackend, EvalType
 from evalscope.evaluator import Evaluator
-from evalscope.models.custom import CustomModel
+from evalscope.models import CustomModel, LocalModel
 from evalscope.utils import seed_everything
 from evalscope.utils.io_utils import OutputsStructure, are_paths_same
 from evalscope.utils.logger import configure_logging, get_logger
@@ -99,22 +99,21 @@ def evaluate_model(task_cfg: TaskConfig, outputs: OutputsStructure) -> dict:
     """Evaluate the model based on the provided task configuration."""
     # Initialize evaluator
     eval_results = {}
-
+    base_model = get_base_model(task_cfg)
     for dataset_name in task_cfg.datasets:
-        evaluator = create_evaluator(task_cfg, dataset_name, outputs)
+        evaluator = create_evaluator(task_cfg, dataset_name, outputs, base_model)
         res_dict = evaluator.eval(infer_cfg=task_cfg.generation_config, debug=task_cfg.debug, limit=task_cfg.limit)
         eval_results[dataset_name] = res_dict
 
     return eval_results
 
 
-def create_evaluator(task_cfg: TaskConfig, dataset_name: str, outputs: OutputsStructure):
+def create_evaluator(task_cfg: TaskConfig, dataset_name: str, outputs: OutputsStructure, base_model: LocalModel):
     """Create an evaluator object for the specified dataset."""
-    # imported_modules = import_module_util(BENCHMARK_PATH_PREFIX, dataset_name, MEMBERS_TO_IMPORT)
     benchmark: BenchmarkMeta = Benchmark.get(dataset_name)
 
     data_adapter = benchmark.get_data_adapter(config=task_cfg.dataset_args)
-    model_adapter = initialize_model_adapter(task_cfg, model_adapter_cls=benchmark.model_adapter)
+    model_adapter = initialize_model_adapter(task_cfg, benchmark.model_adapter, base_model)
 
     return Evaluator(
         dataset_name_or_path=benchmark.dataset_id,
@@ -131,10 +130,31 @@ def create_evaluator(task_cfg: TaskConfig, dataset_name: str, outputs: OutputsSt
     )
 
 
-def initialize_model_adapter(task_cfg: TaskConfig, model_adapter_cls):
+def get_base_model(task_cfg: TaskConfig) -> Optional[LocalModel]:
+    """Get the base model for the task."""
+    if task_cfg.eval_type != EvalType.CHECKPOINT:
+        return None
+    else:
+        device_map = task_cfg.model_args.get('device_map', 'auto') if torch.cuda.is_available() else None
+        cache_dir = task_cfg.model_args.get('cache_dir', None)
+        model_precision = task_cfg.model_args.get('precision', torch.float16)
+        model_revision = task_cfg.model_args.get('revision', DEFAULT_MODEL_REVISION)
+        if isinstance(model_precision, str) and model_precision != 'auto':
+            model_precision = eval(model_precision)
+
+        base_model = LocalModel(
+            model_id=task_cfg.model,
+            model_revision=model_revision,
+            device_map=device_map,
+            torch_dtype=model_precision,
+            cache_dir=cache_dir)
+        return base_model
+
+
+def initialize_model_adapter(task_cfg: TaskConfig, model_adapter_cls, base_model: LocalModel):
     """Initialize the model adapter based on the task configuration."""
     if task_cfg.dry_run:
-        from evalscope.models.dummy_chat_model import DummyChatModel
+        from evalscope.models.model import DummyChatModel
         return DummyChatModel(model_cfg=dict())
     elif task_cfg.eval_type == EvalType.CUSTOM:
         if not isinstance(task_cfg.model, CustomModel):
@@ -142,15 +162,8 @@ def initialize_model_adapter(task_cfg: TaskConfig, model_adapter_cls):
         from evalscope.models.model_adapter import CustomModelAdapter
         return CustomModelAdapter(custom_model=task_cfg.model)
     else:
-        device_map = task_cfg.model_args.get('device_map', 'auto') if torch.cuda.is_available() else None
-        model_precision = task_cfg.model_args.get('precision', torch.float16)
-        if isinstance(model_precision, str) and model_precision != 'auto':
-            model_precision = eval(model_precision)
         return model_adapter_cls(
-            model_id=task_cfg.model,
-            model_revision=task_cfg.model_args.get('revision', DEFAULT_MODEL_REVISION),
-            device_map=device_map,
-            torch_dtype=model_precision,
+            model=base_model or get_base_model(task_cfg),
             generation_config=task_cfg.generation_config,
             chat_template=task_cfg.chat_template)
